WIP: Fusion Parser: Refactor to namespace 'ObjectTreeParser'
rename AstBuilder to better name ObjectTree
